+++ Sources/MapboxNavigationUIKit/Utils/CoreConfig+Conversions.swift
@@ -1,0 +1,15 @@
+import MapboxDirections
+import MapboxNavigationCore
+
+extension CoreConfig {
+    var distanceMeasurementSystem: MeasurementSystem {
+        switch unitOfMeasurement {
+        case .auto:
+            locale.usesMetricSystem ? .metric : .imperial
+        case .metric:
+            .metric
+        case .imperial:
+            .imperial
+        }
+    }
+}